Merge dev branch with staging

+++ resources/views/production/pre-production/_index_filtered.blade.php
@@ -11,14 +11,18 @@
                             <th class="erp-th text-center">Process <span class="erp-tooltip" data-bs-toggle="tooltip" data-bs-placement="top" title="Production Process"><i class="fa-duotone fa-exclamation"></i></span> </th>
                             <th class="erp-th text-center">Raw Materials </th>
                             <th class="erp-th text-center">Estimated QTY </th>
-                            <th class="erp-th text-center">Inventory</th>
-                            <th class="erp-th text-center">Verification </th>
+                            {{-- <th class="erp-th text-center">Inventory</th> --}}
+                            @if(hasPermission('manage-pre-productions'))
+                                <th class="erp-th text-center">Verification </th>
+                            @endif
                             <th class="erp-th text-center">Instruction </th>
-                            <th class="erp-th text-center">Action </th>
+                            @if(hasPermission('manage-pre-productions'))
+                                <th class="erp-th text-center">Action </th>
+                            @endif
                         </tr>
                     </thead>
                     <tbody class="erp-tbody">
-                        @foreach($pre_productions as $data)
+                        @forelse($pre_productions as $data)
                             <tr class="erp-tbody-tr">
                                 <td class="erp-tbody-td">
                                     <h4 class="d-table-title">{{ $pre_productions->firstItem() + $loop->iteration - 1 }}</h4>
@@ -35,11 +39,14 @@
                                     </a>
                                 </td>
                                 <td class="erp-tbody-td text-center">
-                                    <a href="javascript:void(0)" onclick="getDocunent({{$data->id}})" class="document-view-status-btn" data-bs-toggle="modal" data-bs-target="#check_status">
-                                        <img src="{{ asset('assets/img/product/documents.png') }}" alt="" class="document-img-box"><small>View</small>
-                                    </a>
+                                    @if($data->design_of_documents)
+                                        <a href="javascript:void(0)" onclick="getDocunent({{$data->id}})" class="document-view-status-btn" data-bs-toggle="modal" data-bs-target="#check_status">
+                                            <img src="{{ asset('assets/img/product/documents.png') }}" alt="" class="document-img-box"><small>View</small>
+                                        </a>
+                                    @else N/A
+                                    @endif
                                 </td>
-                                
+
                                 <td class="erp-tbody-td text-center">
                                     <h4 class="text-center d-table-title">{{count($data->process)}}</h4>
                                 </td>
@@ -49,44 +56,53 @@
                                 <td class="erp-tbody-td text-center">
                                     <h4 class="text-center d-table-title">{{$data->estimated_production_qty}}</h4>
                                 </td>
-                                <td class="erp-tbody-td text-center">
+                                {{-- <td class="erp-tbody-td text-center">
                                     <a href="#" class="last-cal-status-btn" data-bs-toggle="modal" data-bs-target="#check_in_status">Check Info</a>
-                                </td>
-                                
-                                <td class="erp-tbody-td text-center">
-                                    <div class="checkbox-wrapper">
-                                        <input {{$data->is_verified ==$data::VERIFIED_YES? 'checked disabled': '' }}  id="terms-checkbox-{{$data->id}}" name="checkbox" onclick="updateStatus(this, function () { getData() })" data-href="{{ route('production.pre-production.change-status',[$data->id,1]) }}" type="checkbox">
-                                        <label class="terms-label" for="terms-checkbox-{{$data->id}}">
-                                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 200 200" class="checkbox-svg">
-                                            <mask fill="white" id="path-1-inside-1_476_5-37">
-                                                <rect height="200" width="200"></rect>
-                                            </mask>
-                                            <rect mask="url(#path-1-inside-1_476_5-37)" stroke-width="40" class="checkbox-box" height="200" width="200"></rect>
-                                            <path stroke-width="15" d="M52 111.018L76.9867 136L149 64" class="checkbox-tick"></path>
-                                            </svg>
-                                            <span class="label-text">{{$data->is_verified ==$data::VERIFIED_YES? 'Verified': 'Verify' }}</span>
-                                        </label>
-                                    </div>
-                                </td>
+                                </td> --}}
+                                @if(hasPermission('manage-pre-productions'))
+                                    <td class="erp-tbody-td text-center">
+                                        <div class="checkbox-wrapper">
+                                            <input {{$data->is_verified ==$data::VERIFIED_YES? 'checked disabled': '' }}  id="terms-checkbox-{{$data->id}}" name="checkbox" onclick="updateStatus(this, function () { getData() })" data-href="{{ route('production.pre-production.change-status',[$data->id,1]) }}" type="checkbox">
+                                            <label class="terms-label" for="terms-checkbox-{{$data->id}}">
+                                                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 200 200" class="checkbox-svg">
+                                                <mask fill="white" id="path-1-inside-1_476_5-37">
+                                                    <rect height="200" width="200"></rect>
+                                                </mask>
+                                                <rect mask="url(#path-1-inside-1_476_5-37)" stroke-width="40" class="checkbox-box" height="200" width="200"></rect>
+                                                <path stroke-width="15" d="M52 111.018L76.9867 136L149 64" class="checkbox-tick"></path>
+                                                </svg>
+                                                <span class="label-text">{{$data->is_verified ==$data::VERIFIED_YES? 'Verified': 'Verify' }}</span>
+                                            </label>
+                                        </div>
+                                    </td>
+                                @endif
                                 <td class="erp-tbody-td text-center pre-description-box-td">
                                     <p class="text-center d-table-title pre-description-box">{{$data->description}}</p>
                                 </td>
-                                <td class="text-end erp-tbody-td">
-                                    @if($data->is_verified ==$data::VERIFIED_NO)
-                                        <div class="erp-action-t">
-                                            <div class="dropdown dropdown-action">
-                                                <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
-                                                <div class="dropdown-menu dropdown-menu-right">
-                                                    {{-- <a class="dropdown-item" href="{{ route('production.pre-production.edit',$data->id) }}"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a> --}}
-                                                    <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('production.pre-production.delete',$data->id) }}', 'reloadAjaxGetData') "><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
-                                                    
+                                @if(hasPermission('manage-pre-productions'))
+                                    <td class="text-end erp-tbody-td">
+                                        @if($data->is_verified ==$data::VERIFIED_NO)
+                                            <div class="erp-action-t">
+                                                <div class="dropdown dropdown-action">
+                                                    <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
+                                                    <div class="dropdown-menu dropdown-menu-right">
+                                                        <a class="dropdown-item" href="{{ route('production.pre-production.edit',$data->id) }}"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a>
+                                                        <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('production.pre-production.delete',$data->id) }}', 'reloadAjaxGetData') "><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
+
+                                                    </div>
                                                 </div>
                                             </div>
-                                        </div>
-                                    @endif
+                                        @endif
+                                    </td>
+                                @endif
+                            </tr>
+                        @empty
+                            <tr class="erp-tbody-tr">
+                                <td class="erp-tbody-td text-center text-primary" colspan="10">
+                                    No data found...!
                                 </td>
                             </tr>
-                        @endforeach
+                        @endforelse
                     </tbody>
                 </table>
             </div>
